# api/openapi.yaml
openapi: 3.0.3
info:
  title: Bilibili RAG 知识库系统 API
  description: |
    B站收藏内容可检索知识库系统的 RESTful API 文档。

    ## 核心能力

    1. **数据获取** - B站扫码登录、收藏夹/视频列表同步
    2. **内容提取** - ASR（语音转文本）、基础信息（标题/简介）
    3. **知识构建** - 文本切块（chunking）、向量化（embedding）、存储到 ChromaDB
    4. **智能问答（RAG）** - 查询路由（direct/db/vector）、向量检索+LLM生成、流式返回+来源追踪

    ## 认证机制

    - 使用 `session_id` 作为会话标识
    - 扫码登录后获取 `session_id`
    - 大部分接口需要携带 `session_id` 参数

    ## SSE 流式响应

    聊天接口 `/chat/ask/stream` 使用 Server-Sent Events 流式返回：
    - 格式: `data: {chunk_text}`
    - 结束后发送: `data: [[SOURCES_JSON]]{sources_json}`
  version: 1.2.0
  contact:
    name: Bilibili RAG Project
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
...
paths:
...
  /favorites/list:
    get:
      tags:
        - 收藏夹
      summary: 获取收藏夹列表
      description: |
        获取当前登录用户的所有收藏夹列表。

        ### 返回字段说明

        - `media_id`: 收藏夹ID
        - `title`: 收藏夹名称
        - `media_count`: 收藏夹内视频数量
        - `is_selected`: 是否选中（默认 true）
        - `is_default`: 是否默认收藏夹
      operationId: getFavoritesList
      parameters:
        - name: session_id
          in: query
          required: true
          description: 会话ID
          schema:
            type: string
      responses:
        '200':
          description: 收藏夹列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FavoriteFolderInfo'
              example:
                - media_id: 123456
                  title: "学习资料"
                  media_count: 42
                  is_selected: true
                  is_default: false
                - media_id: 789012
                  title: "默认收藏夹"
                  media_count: 15
                  is_selected: true
                  is_default: true
        '401':
          description: 未登录或会话已过期
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /favorites/{media_id}/videos:
    get:
      tags:
        - 收藏夹
      summary: 获取收藏夹视频（分页）
      description: |
        获取指定收藏夹中的视频列表，支持分页。

        ### 分页参数

        - `page`: 页码，从 1 开始
        - `page_size`: 每页数量，默认 20，最大 20
      operationId: getFavoriteVideos
      parameters:
        - name: media_id
          in: path
          required: true
          description: 收藏夹媒体ID
          schema:
            type: integer
        - name: session_id
          in: query
          required: true
          description: 会话ID
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: 页码
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          required: false
          description: 每页数量
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 20
      responses:
        '200':
          description: 收藏夹视频列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteVideosResponse'
              example:
                folder_info:
                  title: "学习资料"
                  media_count: 42
                videos:
                  - bvid: "BV1xx411c7mD"
                    title: "Python 入门教程"
                    cover: "https://i0.hdslb.com/bfs/archive/xxx.jpg"
                    duration: "12:34"
                    owner: "UP主名称"
                    play_count: 12345
                    intro: "Python 编程入门..."
                    is_selected: true
                has_more: true
                page: 1
                page_size: 20
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /favorites/{media_id}/all-videos:
    get:
      tags:
        - 收藏夹
      summary: 获取收藏夹所有视频
      description: |
        获取指定收藏夹中的所有视频（用于构建知识库）。

        ### 注意事项

        - 自动过滤失效视频（已删除、已下架等）
        - 返回所有有效视频信息
        - 包含 `cid` 字段，用于后续 ASR 处理
      operationId: getAllFavoriteVideos
      parameters:
        - name: media_id
          in: path
          required: true
          description: 收藏夹媒体ID
          schema:
            type: integer
        - name: session_id
          in: query
          required: true
          description: 会话ID
          schema:
            type: string
      responses:
        '200':
          description: 所有视频列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    description: 有效视频总数
                  videos:
                    type: array
                    items:
                      $ref: '#/components/schemas/VideoInfo'
              example:
                total: 42
                videos:
                  - bvid: "BV1xx411c7mD"
                    title: "Python 入门教程"
                    cover: "https://i0.hdslb.com/bfs/archive/xxx.jpg"
                    duration: "12:34"
                    owner: "UP主名称"
                    cid: 12345678
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /favorites/organize/preview:
    post:
      tags:
        - 收藏夹
      summary: 预览收藏夹整理
      description: |
        预览按已有收藏夹名称对默认收藏夹内容分类的结果。

        ### 功能说明

        - 获取默认收藏夹中的所有视频
        - 按已有收藏夹名称进行匹配
        - 返回预览结果供用户确认
      operationId: organizePreview
      parameters:
        - name: session_id
          in: query
          required: true
          description: 会话ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizePreviewRequest'
      responses:
        '200':
          description: 预览结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizePreviewResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /favorites/organize/execute:
    post:
      tags:
        - 收藏夹
      summary: 执行收藏夹整理
      description: |
        根据预览结果批量移动收藏夹内容。

        ### 请求格式

        - `default_folder_id`: 默认收藏夹ID（源）
        - `moves`: 移动操作列表

        ### 移动格式

        ```json
        {
          "resource_id": 123,
          "resource_type": 2,
          "target_folder_id": 456
        }
        ```
      operationId: organizeExecute
      parameters:
        - name: session_id
          in: query
          required: true
          description: 会话ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizeExecuteRequest'
      responses:
        '200':
          description: 执行结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  moved:
                    type: integer
                  groups:
                    type: integer
              example:
                message: "移动完成"
                moved: 15
                groups: 3
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /favorites/organize/clean-invalid:
    post:
      tags:
        - 收藏夹
      summary: 清理失效内容
      description: |
        清理指定收藏夹中的失效内容（已删除、已下架视频）。
      operationId: cleanInvalidResources
      parameters:
        - name: session_id
          in: query
          required: true
          description: 会话ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CleanInvalidRequest'
      responses:
        '200':
          description: 清理完成
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
              example:
                message: "清理完成"
                data:
                  deleted: 5

  # ========== 知识库路由 ==========
...
components:
  schemas:
...
    FavoriteFolderInfo:
      type: object
      properties:
        media_id:
          type: integer
          description: 收藏夹ID
        title:
          type: string
          description: 收藏夹标题
        media_count:
          type: integer
          description: 视频数量
        is_selected:
          type: boolean
          description: 是否选中
        is_default:
          type: boolean
          description: 是否默认收藏夹
    FavoriteVideosResponse:
      type: object
      properties:
        folder_info:
          type: object
          properties:
            title:
              type: string
            media_count:
              type: integer
        videos:
          type: array
          items:
            $ref: '#/components/schemas/FavoriteVideoItem'
        has_more:
          type: boolean
        page:
          type: integer
        page_size:
          type: integer
    FavoriteVideoItem:
      type: object
      properties:
        bvid:
          type: string
        title:
          type: string
        cover:
          type: string
        duration:
          type: string
        owner:
          type: string
        play_count:
          type: integer
        intro:
          type: string
        is_selected:
          type: boolean
    VideoInfo:
      type: object
      properties:
        bvid:
          type: string
        title:
          type: string
        cover:
          type: string
        duration:
          type: string
        owner:
          type: string
        cid:
          type: integer
          description: 视频CID（用于ASR）
    OrganizePreviewRequest:
      type: object
      required:
        - folder_id
      properties:
        folder_id:
          type: integer
          description: 默认收藏夹ID
    OrganizePreviewResponse:
      type: object
      properties:
        default_folder_id:
          type: integer
        default_folder_title:
          type: string
        folders:
          type: array
          items:
            $ref: '#/components/schemas/FavoriteFolderInfo'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrganizePreviewItem'
        stats:
          type: object
          properties:
            total:
              type: integer
            matched:
              type: integer
            unmatched:
              type: integer
    OrganizePreviewItem:
      type: object
      properties:
        bvid:
          type: string
        title:
          type: string
        resource_id:
          type: integer
        resource_type:
          type: integer
        target_folder_id:
          type: integer
          nullable: true
        target_folder_title:
          type: string
        reason:
          type: string
          nullable: true
    OrganizeExecuteRequest:
      type: object
      required:
        - default_folder_id
        - moves
      properties:
        default_folder_id:
          type: integer
        moves:
          type: array
          items:
            $ref: '#/components/schemas/OrganizeMoveItem'
    OrganizeMoveItem:
      type: object
      properties:
        resource_id:
          type: integer
        resource_type:
          type: integer
        target_folder_id:
          type: integer
    CleanInvalidRequest:
      type: object
      required:
        - folder_id
      properties:
        folder_id:
          type: integer

    # ========== 知识库相关 ==========
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: 错误详情